moved config classes to subfolder in tool
